split createContainer
