+++ package.json
@@ -23,9 +23,7 @@
     "eslint": "eslint --ext .js,.ts,.tsx --ignore-pattern dist .",
     "jest": "jest",
     "tsc-test": "tsc --project . --noEmit",
-    "preapidoc": "sed 's/ as const;/;/g' src/createStore.ts > src/createStore_for_doc.ts",
-    "apidoc": "documentation readme --section API --markdown-toc false --parse-extension ts src/createGlobalState.ts src/createStore_for_doc.ts",
-    "postapidoc": "rm src/createStore_for_doc.ts",
+    "apidoc": "documentation readme --section API --markdown-toc false --parse-extension ts src/createGlobalState.ts src/createStore.ts",
     "examples:01_minimal": "DIR=01_minimal EXT=js webpack-dev-server",
     "examples:02_typescript": "DIR=02_typescript webpack-dev-server",
     "examples:03_actions": "DIR=03_actions webpack-dev-server",
